Refactor resize package API
- resizing can now be done in a single call without managing state.
  e.g: Rather than creating a new 'resizer': resize.DefaultResize()
  or something similar, you can call
    resize.Buffer(someBytes, option)
  where someBytes is an image buffer, and option is currently one of
    - ToScale(scale float64)
    - ToFixed(width, height int)
    - ToTerminal()

  This allows the user to resize using predefined options without
  faffing with values and error checking, since the check is done
  at compile time, and options are constrained by function signatures.

- Add vips_resize.go which contains resizing functions that use vips
bindings (for speed?). This means vips is currently required to run
the program, so the 'vips' build flag has been added, and another file
will be added that uses a native Go library to resize.
This means two versions will eventually be produced, one with vips
bindings, and one without.